requirements.txt
transformers==4.56.1
modelscope>=1.23
peft>=0.11,<0.18
trl>=0.15,<0.21
deepspeed>=0.14
vllm>=0.5.1
sglang>=0.4.6
lmdeploy>=0.5
evalscope>=0.11
gradio
ms-swift
qwen_vl_utils[decord]
json-repair
mpi4py